requirements.txt
Flask~=2.0.1
pytest
Werkzeug~=2.0.2
SQLAlchemy~=1.4.25
pg8000==1.21.3
pyjwt==2.2.0